Rename 'full' to 'airport'
